pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "eurosat-land-classification"
version = "0.1.0"
description = "EuroSAT land-use classification project scaffold"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Milos Milanovic" },
  { name = "Bojan Zivanic" },
]
dependencies = [
  "torch>=2.5.0",
  "torchvision>=0.20.0",
  "Pillow>=10.4.0",
  "numpy>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.11.0",
  "mypy>=1.11.0",
  "types-Pillow>=10.2.0.20240822",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E9", "F63", "F7", "F82"]

[tool.mypy]
python_version = "3.12"
files = ["src"]
ignore_missing_imports = true
warn_unused_ignores = true
show_error_codes = true
